added checks to show/hide HandInItem button depending on if player has valid quest for item. players can hand in items they want and quests auto complete. added checks based on amount of items needed to hand in and items current stack count. EG: handing in 3 potions when u need 5 wont auto complete quest. handing in 3 potions when u only need 2 will leave u with 1 potion etc...

+++ Assets/Scripts/Managers/SaveManager.cs
@@ -356,8 +356,8 @@ public class QuestItemData
 	{
 		isBossKillQuest, isKillQuest, isItemHandInQuest
 	}
-	public float amount;
-	public float currentAmount;
+	public int amount;
+	public int currentAmount;
 
 	[Header("Kill Quest Info")]
 	public SOEntityStats entityToKill;
